requirements.txt
dash==2.6.0
dash_bootstrap_components==1.1.0
Flask==2.1.2
keras==2.9.0
numpy==1.21.6
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_gpu==2.9.0
